[MSE][GStreamer] Fix the deadlock caused by bug 185242 (r231351) https://bugs.webkit.org/show_bug.cgi?id=185723
Patch by Yacine Bandou <yacine.bandou_ext@softathome.com> on 2018-05-22
Reviewed by Xabier Rodriguez-Calvar.

After a detailed investigation, we found that, the patch r231351 doesn't fix the crash that expected to fix,
it just replaces it by a deadlock. Now the crash is fixed in the bug 185725.

This Patch will remove a part of r231351 that causes the deadlock.

* platform/graphics/gstreamer/mse/WebKitMediaSourceGStreamer.cpp:
(webKitMediaSrcFreeStream):

Identifier: 201309@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232058 268f45cc-cd09-0410-ab3c-d52691b4dbfc