AX: setValue on contenteditable should preserve whitespace https://bugs.webkit.org/show_bug.cgi?id=185897
Reviewed by Chris Fleizach.

Source/WebCore:

RenderText is using its parent renderer's style to determine if
whitespace collapsing is necessary. So when setting the innerText
of the element in setValue, let's also set its style in order to
preserve whitespaces.

Modified an existing test to cover this change.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::setValue):

LayoutTests:

* accessibility/mac/set-value-editable-types-expected.txt:
* accessibility/mac/set-value-editable-types.html:


Identifier: 201363@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232120 268f45cc-cd09-0410-ab3c-d52691b4dbfc